+++ src/services/ProductService.ts
@@ -1,1 +1,10 @@
-export default class ProductService {}
+export default class ProductService {
+  public static async getProducts(): Promise<object[]> {
+    return fetch("http://localhost:8080/api/products").then((response) =>
+      response.json()
+    )
+    .catch(err => 
+        console.log(err)
+    );
+  }
+}
